remove unnecessary cast

+++ scipy/optimize/_lsap.py
@@ -97,4 +97,4 @@ def linear_sum_assignment(cost_matrix, maximize=False):
     if maximize:
         cost_matrix = -cost_matrix
 
-    return _lsap_module.calculate_assignment(cost_matrix.astype(np.double))
+    return _lsap_module.calculate_assignment(cost_matrix)
